requirements.txt
openai==1.30.1
python-telegram-bot==20.7
python-dotenv==1.0.1
schedule==1.2.1
